# docs/swagger.yaml
basePath: /api
definitions:
  models.Applicant:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      employment_status:
        type: string
      household:
        items:
          $ref: '#/definitions/models.HouseholdMember'
        type: array
      id:
        type: string
      marital_status:
        type: string
      name:
        type: string
      sex:
        type: string
      updated_at:
        type: string
    type: object
  models.ApplicantResponse:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      employment_status:
        type: string
      household:
        items:
          $ref: '#/definitions/models.HouseholdMember'
        type: array
      id:
        type: string
      marital_status:
        type: string
      name:
        type: string
      sex:
        type: string
      updated_at:
        type: string
    type: object
...
  models.HouseholdMember:
    properties:
      applicant_id:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      employment_status:
        type: string
      id:
        type: string
      name:
        type: string
      relation:
        type: string
      sex:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
paths:
  /api/applicants:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all applicants with their household members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ApplicantResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all applicants
      tags:
      - applicants
    post:
      consumes:
      - application/json
      description: Add a new applicant to the system
      parameters:
      - description: Applicant information
        in: body
        name: applicant
        required: true
        schema:
          $ref: '#/definitions/models.Applicant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ApplicantResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new applicant
      tags:
      - applicants
  /api/applicants/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an applicant from the system
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Applicant not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete applicant
      tags:
      - applicants
    get:
      consumes:
      - application/json
      description: Retrieve a specific applicant by their ID
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApplicantResponse'
        "404":
          description: Applicant not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get applicant by ID
      tags:
      - applicants
    put:
      consumes:
      - application/json
      description: Update an existing applicant's information
      parameters:
      - description: Applicant ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated applicant information
        in: body
        name: applicant
        required: true
        schema:
          $ref: '#/definitions/models.Applicant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Applicant'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Applicant not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update applicant
      tags:
      - applicants
...
schemes:
- http
swagger: "2.0"